# docs/contracts/openapi-v1.json
{
  "openapi": "3.1.0",
...
  "info": {"title": "StudyPilot 第一阶段 API", "version": "1.0.0", "description": "本地单用户 StudyPilot 第一阶段完整 MVP 目标契约，不是当前全部可调用能力清单。TASK-009 的已授权阶段子集及暂未开放输入响应见 x-delivery-profile 和 API与数据契约基线.md 第1.3节；当前运行时可用性以该阶段说明为准。标准 paths/schemas 保留最终要求，不能未经阶段筛选直接生成当前客户端。事务、安全前置顺序、文件对账与删除恢复的含义权威见同一契约文档。"},
...
  "paths": {
...
    "/api/v1/resources": {
      "parameters": [{"$ref": "#/components/parameters/HostHeader"}],
      "get": {
        "operationId": "listResources",
        "tags": ["resources"],
        "summary": "分页搜索和筛选资料",
        "description": "默认排除 ARCHIVED；明确 learning_status=ARCHIVED 可查看。搜索仅覆盖标题、来源名称、保存原因。见契约第2.3、5节。",
        "x-contract-section": "2.3,5",
        "security": [{"LocalToken": []}],
        "parameters": [{"$ref": "#/components/parameters/Page"}, {"$ref": "#/components/parameters/PageSize"}, {"name": "q", "in": "query", "schema": {"type": "string", "minLength": 1, "maxLength": 200}}, {"name": "topic_id", "in": "query", "schema": {"type": "string", "format": "uuid"}}, {"name": "topic_unassigned", "in": "query", "schema": {"type": "boolean"}}, {"name": "tag_id", "in": "query", "description": "可重复；必须同时具有全部标签。", "schema": {"type": "array", "items": {"type": "string", "format": "uuid"}}}, {"name": "source_type", "in": "query", "schema": {"type": "array", "items": {"$ref": "#/components/schemas/ResourceSourceType"}}}, {"name": "learning_status", "in": "query", "schema": {"type": "array", "items": {"$ref": "#/components/schemas/LearningStatus"}}}, {"name": "progress_min", "in": "query", "schema": {"type": "integer", "minimum": 0, "maximum": 100}}, {"name": "progress_max", "in": "query", "schema": {"type": "integer", "minimum": 0, "maximum": 100}}, {"name": "created_from", "in": "query", "schema": {"type": "string", "format": "date-time"}}, {"name": "created_to", "in": "query", "schema": {"type": "string", "format": "date-time"}}, {"name": "updated_from", "in": "query", "schema": {"type": "string", "format": "date-time"}}, {"name": "updated_to", "in": "query", "schema": {"type": "string", "format": "date-time"}}, {"name": "sort", "in": "query", "schema": {"type": "string", "enum": ["created_at", "-created_at", "updated_at", "-updated_at", "title", "-title", "progress_percent", "-progress_percent"], "default": "-created_at"}}],
        "responses": {"200": {"description": "稳定分页资料摘要。", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/ResourcePage"}, "example": {"data": [], "page": {"number": 1, "size": 20, "total_items": 0, "total_pages": 0, "has_more": false}}}}}, "403": {"$ref": "#/components/responses/ForbiddenRead"}, "422": {"$ref": "#/components/responses/Validation"}, "500": {"$ref": "#/components/responses/ServerError"}},
        "x-error-codes": ["HOST_FORBIDDEN", "LOCAL_TOKEN_REQUIRED", "LOCAL_TOKEN_INVALID", "VALIDATION_ERROR", "UNKNOWN_ERROR"]
      },
      "post": {
        "operationId": "createResource",
        "tags": ["resources"],
        "summary": "添加网页、文件或粘贴资料",
        "description": "完整目标：JSON 使用 WEB/PASTE oneOf；multipart 只用于 FILE，文件达到 READY 才成功。阶段可用性以契约第1.3节及 x-delivery-profile 为准；最终要求见第5、7、8节。",
        "x-contract-section": "5,7,8",
        "security": [{"LocalToken": []}],
        "parameters": [{"$ref": "#/components/parameters/OriginHeader"}, {"$ref": "#/components/parameters/SecFetchSite"}, {"$ref": "#/components/parameters/SecFetchMode"}, {"$ref": "#/components/parameters/SecFetchDest"}],
        "requestBody": {"required": true, "content": {"application/json": {"schema": {"oneOf": [{"$ref": "#/components/schemas/WebResourceCreate"}, {"$ref": "#/components/schemas/PasteResourceCreate"}], "discriminator": {"propertyName": "source_type"}}, "examples": {"web": {"value": {"source_type": "WEB", "title": "示例文档", "source_url": "https://example.test/guide", "source_name": "示例站点", "save_reason": "学习路由", "topic_id": null, "tag_ids": []}}, "paste": {"value": {"source_type": "PASTE", "title": "个人摘录", "pasted_content": "# 示例内容", "topic_id": null, "tag_ids": []}}}}, "multipart/form-data": {"schema": {"$ref": "#/components/schemas/FileResourceCreate"}, "example": {"source_type": "FILE", "title": "示例讲义", "file": "<binary-file-content-omitted>", "source_name": "本地课程", "save_reason": "保存原始讲义", "topic_id": null, "tag_ids": []}}}},
        "responses": {"201": {"description": "完整资料；FILE 必为 READY。", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/ResourceEnvelope"}, "example": {"data": {"id": "018f1f58-4eb2-4a0d-a716-fb81b1960001", "title": "示例资料", "source_type": "WEB", "source_name": "示例站点", "save_reason": "学习接口设计", "topic_id": null, "version": 1, "created_at": "2026-09-02T02:00:00Z", "updated_at": "2026-09-02T02:00:00Z", "progress": {"resource_id": "018f1f58-4eb2-4a0d-a716-fb81b1960001", "status": "UNREAD", "progress_percent": 0, "started_at": null, "completed_at": null, "archived_from_status": null, "version": 1, "updated_at": "2026-09-02T02:00:00Z"}, "tags": [], "original_file": null, "review_plan": null, "source_url": "https://example.test/guide"}}}}}, "400": {"$ref": "#/components/responses/BadRequest"}, "403": {"$ref": "#/components/responses/ForbiddenWrite"}, "404": {"$ref": "#/components/responses/TopicOrTagNotFound"}, "413": {"$ref": "#/components/responses/TooLarge"}, "415": {"$ref": "#/components/responses/UnsupportedMedia"}, "422": {"$ref": "#/components/responses/Validation"}, "500": {"$ref": "#/components/responses/ServerError"}, "503": {"$ref": "#/components/responses/StorageUnavailable"}},
        "x-error-codes": ["HOST_FORBIDDEN", "LOCAL_TOKEN_REQUIRED", "LOCAL_TOKEN_INVALID", "REQUEST_ORIGIN_FORBIDDEN", "MALFORMED_REQUEST", "TOPIC_NOT_FOUND", "TAG_NOT_FOUND", "FILE_TOO_LARGE", "CONTENT_TYPE_UNSUPPORTED", "FILE_TYPE_UNSUPPORTED", "VALIDATION_ERROR", "STORAGE_PATH_UNAVAILABLE", "UNKNOWN_ERROR"]
      }
    },
    "/api/v1/resources/{resource_id}": {
      "parameters": [{"$ref": "#/components/parameters/HostHeader"}, {"name": "resource_id", "in": "path", "required": true, "schema": {"type": "string", "format": "uuid"}}],
      "get": {"operationId": "getResource", "tags": ["resources"], "summary": "查看资料详情", "description": "PASTE 可返回原始粘贴内容；FILE 只返回受控元数据。", "x-contract-section": "4.1,5", "security": [{"LocalToken": []}], "responses": {"200": {"description": "资料详情。", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/ResourceEnvelope"}, "example": {"data": {"id": "018f1f58-4eb2-4a0d-a716-fb81b1960001", "title": "示例资料", "source_type": "WEB", "source_name": "示例站点", "save_reason": "学习接口设计", "topic_id": null, "version": 1, "created_at": "2026-09-02T02:00:00Z", "updated_at": "2026-09-02T02:00:00Z", "progress": {"resource_id": "018f1f58-4eb2-4a0d-a716-fb81b1960001", "status": "UNREAD", "progress_percent": 0, "started_at": null, "completed_at": null, "archived_from_status": null, "version": 1, "updated_at": "2026-09-02T02:00:00Z"}, "tags": [], "original_file": null, "review_plan": null, "source_url": "https://example.test/guide"}}}}}, "403": {"$ref": "#/components/responses/ForbiddenRead"}, "404": {"$ref": "#/components/responses/ResourceNotFound"}, "500": {"$ref": "#/components/responses/ServerError"}}, "x-error-codes": ["HOST_FORBIDDEN", "LOCAL_TOKEN_REQUIRED", "LOCAL_TOKEN_INVALID", "RESOURCE_NOT_FOUND", "UNKNOWN_ERROR"]},
      "patch": {
        "operationId": "updateResource",
        "tags": ["resources"],
        "summary": "修改资料元数据或同类型来源内容",
        "description": "省略不修改，允许字段的 null 清空；类型与文件原件不可改。source_url 与 pasted_content 同时出现返回 422 VALIDATION_ERROR；单一来源字段与现有 source_type 不匹配返回 409 SOURCE_TYPE_MISMATCH。",
        "x-contract-section": "2.4,4.1,5",
        "security": [{"LocalToken": []}],
        "parameters": [{"$ref": "#/components/parameters/OriginHeader"}, {"$ref": "#/components/parameters/SecFetchSite"}, {"$ref": "#/components/parameters/SecFetchMode"}, {"$ref": "#/components/parameters/SecFetchDest"}],
        "requestBody": {"required": true, "content": {"application/json": {"schema": {"$ref": "#/components/schemas/ResourcePatch"}, "example": {"expected_version": 1, "save_reason": "补充实践原因"}}}},
        "responses": {"200": {"description": "更新后的详情。", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/ResourceEnvelope"}, "example": {"data": {"id": "018f1f58-4eb2-4a0d-a716-fb81b1960001", "title": "示例资料", "source_type": "WEB", "source_name": "示例站点", "save_reason": "补充实践原因", "topic_id": null, "version": 2, "created_at": "2026-09-02T02:00:00Z", "updated_at": "2026-09-02T02:05:00Z", "progress": {"resource_id": "018f1f58-4eb2-4a0d-a716-fb81b1960001", "status": "UNREAD", "progress_percent": 0, "started_at": null, "completed_at": null, "archived_from_status": null, "version": 1, "updated_at": "2026-09-02T02:00:00Z"}, "tags": [], "original_file": null, "review_plan": null, "source_url": "https://example.test/guide"}}}}}, "400": {"$ref": "#/components/responses/BadRequest"}, "403": {"$ref": "#/components/responses/ForbiddenWrite"}, "404": {"$ref": "#/components/responses/ResourceOrTopicNotFound"}, "409": {"$ref": "#/components/responses/ResourceUpdateConflict"}, "415": {"$ref": "#/components/responses/UnsupportedContentType"}, "422": {"$ref": "#/components/responses/Validation"}, "428": {"$ref": "#/components/responses/PreconditionRequired"}, "500": {"$ref": "#/components/responses/ServerError"}},
        "x-error-codes": ["HOST_FORBIDDEN", "LOCAL_TOKEN_REQUIRED", "LOCAL_TOKEN_INVALID", "REQUEST_ORIGIN_FORBIDDEN", "MALFORMED_REQUEST", "RESOURCE_NOT_FOUND", "TOPIC_NOT_FOUND", "VERSION_CONFLICT", "SOURCE_TYPE_MISMATCH", "CONTENT_TYPE_UNSUPPORTED", "VALIDATION_ERROR", "VERSION_REQUIRED", "UNKNOWN_ERROR"]
      },
      "delete": {
        "operationId": "deleteResource",
        "tags": ["resources"],
        "summary": "使用一次性确认令牌删除资料",
        "description": "事务内重算稳定影响集合；变化时旧令牌作废，409 只返回当前影响摘要，用户重新调用 deletion-preview 才能取得新令牌。文件先移入同卷 trash。见契约第9节。",
        "x-contract-section": "9",
        "security": [{"LocalToken": []}],
        "parameters": [{"$ref": "#/components/parameters/OriginHeader"}, {"$ref": "#/components/parameters/SecFetchSite"}, {"$ref": "#/components/parameters/SecFetchMode"}, {"$ref": "#/components/parameters/SecFetchDest"}, {"$ref": "#/components/parameters/DeletionToken"}],
        "responses": {"204": {"description": "删除事务已提交；没有响应体。"}, "403": {"$ref": "#/components/responses/ForbiddenWrite"}, "404": {"$ref": "#/components/responses/ResourceNotFound"}, "409": {"description": "令牌已重放，或影响集合变化。变化时旧令牌作废且错误响应绝不返回新令牌。", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/DeletionConflictErrorResponse"}, "examples": {"impactChanged": {"value": {"error": {"code": "DELETION_IMPACT_CHANGED", "message": "删除影响已经变化，请重新确认。", "details": {"current_impact": {"resource_id": "018f1f58-4eb2-4a0d-a716-fb81b1960001", "resource_version": 4, "impact_revision": "bbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbb", "impact": {"original_file_count": 1, "note_count": 3, "study_record_count": 4, "active_review_plan_count": 1, "review_record_count": 1, "resource_tag_count": 2}}}, "request_id": "req_7cc640f6b1764ed0"}}}, "replayed": {"value": {"error": {"code": "DELETION_TOKEN_REPLAYED", "message": "这次删除确认已经使用，请重新预览。", "details": {}, "request_id": "req_7cc640f6b1764ed0"}}}}}}}, "410": {"$ref": "#/components/responses/Gone"}, "500": {"$ref": "#/components/responses/ServerError"}, "503": {"$ref": "#/components/responses/StorageUnavailable"}},
        "x-error-codes": ["HOST_FORBIDDEN", "LOCAL_TOKEN_REQUIRED", "LOCAL_TOKEN_INVALID", "REQUEST_ORIGIN_FORBIDDEN", "RESOURCE_NOT_FOUND", "DELETION_TOKEN_REQUIRED", "DELETION_TOKEN_INVALID", "DELETION_TOKEN_REPLAYED", "DELETION_IMPACT_CHANGED", "DELETION_TOKEN_EXPIRED", "STORAGE_PATH_UNAVAILABLE", "UNKNOWN_ERROR"]
      }
    },
...
  },
  "components": {
    "securitySchemes": {"LocalToken": {"type": "apiKey", "in": "header", "name": "X-StudyPilot-Token", "description": "每次后端启动生成的至少 256 位随机令牌；只保存在前端内存。"}},
    "parameters": {
      "HostHeader": {"name": "Host", "in": "header", "required": true, "description": "必须精确等于配置的可信 authority；重复或转发覆盖值拒绝。", "schema": {"type": "string", "example": "127.0.0.1:8000"}},
      "OriginHeader": {"name": "Origin", "in": "header", "required": true, "description": "写请求必须精确等于规范 UI 来源。", "schema": {"type": "string", "format": "uri", "example": "http://127.0.0.1:5173"}},
...
      "SecFetchSite": {"name": "Sec-Fetch-Site", "in": "header", "required": true, "schema": {"type": "string", "const": "same-origin"}},
      "SecFetchMode": {"name": "Sec-Fetch-Mode", "in": "header", "required": true, "schema": {"type": "string", "const": "cors"}},
      "SecFetchDest": {"name": "Sec-Fetch-Dest", "in": "header", "required": true, "schema": {"type": "string", "const": "empty"}},
      "DeletionToken": {"name": "X-StudyPilot-Deletion-Token", "in": "header", "required": true, "description": "删除预览返回的短时一次性不透明令牌。", "schema": {"type": "string", "minLength": 43, "writeOnly": true}},
...
      "Page": {"name": "page", "in": "query", "required": false, "schema": {"type": "integer", "minimum": 1, "default": 1}},
      "PageSize": {"name": "page_size", "in": "query", "required": false, "schema": {"type": "integer", "minimum": 1, "maximum": 100, "default": 20}},
...
    },
    "schemas": {
      "ErrorResponse": {"type": "object", "additionalProperties": false, "required": ["error"], "properties": {"error": {"type": "object", "additionalProperties": false, "required": ["code", "message", "details", "request_id"], "properties": {"code": {"type": "string", "enum": ["MALFORMED_REQUEST", "HOST_FORBIDDEN", "LOCAL_TOKEN_REQUIRED", "LOCAL_TOKEN_INVALID", "REQUEST_ORIGIN_FORBIDDEN", "DELETION_TOKEN_REQUIRED", "DELETION_TOKEN_INVALID", "RESOURCE_NOT_FOUND", "TOPIC_NOT_FOUND", "TAG_NOT_FOUND", "NOTE_NOT_FOUND", "FILE_NOT_FOUND", "DUPLICATE_TOPIC", "DUPLICATE_TAG", "VERSION_CONFLICT", "INVALID_STATE_TRANSITION", "STATE_CONFLICT", "SOURCE_TYPE_MISMATCH", "TAXONOMY_IN_USE", "FILE_STATE_UNAVAILABLE", "FILE_CORRUPTED", "DELETION_TOKEN_REPLAYED", "DELETION_IMPACT_CHANGED", "DELETION_TOKEN_EXPIRED", "FILE_TOO_LARGE", "CONTENT_TYPE_UNSUPPORTED", "FILE_TYPE_UNSUPPORTED", "VALIDATION_ERROR", "VERSION_REQUIRED", "UNKNOWN_ERROR", "STORAGE_PATH_UNAVAILABLE"]}, "message": {"type": "string"}, "details": {"type": "object", "additionalProperties": true}, "request_id": {"type": "string", "pattern": "^req_[A-Za-z0-9_-]+$"}}}}},
      "PageMeta": {"type": "object", "additionalProperties": false, "required": ["number", "size", "total_items", "total_pages", "has_more"], "properties": {"number": {"type": "integer", "minimum": 1}, "size": {"type": "integer", "minimum": 1, "maximum": 100}, "total_items": {"type": "integer", "minimum": 0}, "total_pages": {"type": "integer", "minimum": 0}, "has_more": {"type": "boolean"}}},
      "ResourceSourceType": {"type": "string", "enum": ["WEB", "FILE", "PASTE"], "x-labels-zh-CN": {"WEB": "网页", "FILE": "文件", "PASTE": "粘贴内容"}},
      "LearningStatus": {"type": "string", "enum": ["UNREAD", "IN_PROGRESS", "COMPLETED", "REVIEW_DUE", "ARCHIVED"], "x-labels-zh-CN": {"UNREAD": "未读", "IN_PROGRESS": "学习中", "COMPLETED": "已完成", "REVIEW_DUE": "待复习", "ARCHIVED": "已归档"}},
...
      "SourceUrl": {"type": "string", "format": "uri", "pattern": "^https?://(?![^/?#]*@)[^#]+$", "maxLength": 2048, "x-sensitive": true, "description": "必须解析为 http/https URL；username、password 和 fragment 必须为空。"},
      "LearningResource": {"type": "object", "required": ["id", "title", "source_type", "source_name", "save_reason", "topic_id", "version", "created_at", "updated_at"], "properties": {"id": {"type": "string", "format": "uuid", "readOnly": true}, "title": {"type": "string", "minLength": 1, "maxLength": 200}, "source_type": {"$ref": "#/components/schemas/ResourceSourceType"}, "source_name": {"type": ["string", "null"], "maxLength": 120, "x-sensitive": true}, "save_reason": {"type": ["string", "null"], "maxLength": 1000, "x-sensitive": true}, "topic_id": {"type": ["string", "null"], "format": "uuid"}, "version": {"type": "integer", "minimum": 1, "readOnly": true}, "created_at": {"type": "string", "format": "date-time", "readOnly": true}, "updated_at": {"type": "string", "format": "date-time", "readOnly": true}}},
...
      "DeletionImpactChangedDetails": {"type": "object", "additionalProperties": false, "required": ["current_impact"], "properties": {"current_impact": {"$ref": "#/components/schemas/DeletionImpactCurrent"}}},
      "DeletionImpactChangedErrorResponse": {"description": "统一 ErrorResponse 在 DELETION_IMPACT_CHANGED 场景的收窄形状；机器 schema 不允许确认令牌或过期时间。", "allOf": [{"$ref": "#/components/schemas/ErrorResponse"}, {"type": "object", "properties": {"error": {"type": "object", "required": ["code", "details"], "properties": {"code": {"const": "DELETION_IMPACT_CHANGED"}, "details": {"$ref": "#/components/schemas/DeletionImpactChangedDetails"}}}}}]},
      "DeletionTokenReplayedErrorResponse": {"description": "统一 ErrorResponse 在删除令牌重放场景的收窄形状。", "allOf": [{"$ref": "#/components/schemas/ErrorResponse"}, {"type": "object", "properties": {"error": {"type": "object", "required": ["code", "details"], "properties": {"code": {"const": "DELETION_TOKEN_REPLAYED"}, "details": {"type": "object", "maxProperties": 0}}}}}]},
      "DeletionConflictErrorResponse": {"oneOf": [{"$ref": "#/components/schemas/DeletionImpactChangedErrorResponse"}, {"$ref": "#/components/schemas/DeletionTokenReplayedErrorResponse"}]},
...
      "OriginalFileSummary": {"description": "列表安全文件摘要；不含 sha256、resource_id、failure_code、version 或时间字段。", "type": "object", "additionalProperties": false, "required": ["id", "original_name", "size_bytes", "media_type", "status"], "properties": {"id": {"type": "string", "format": "uuid", "readOnly": true}, "original_name": {"type": "string", "minLength": 1, "maxLength": 255, "readOnly": true, "x-sensitive": true}, "size_bytes": {"type": "integer", "minimum": 1, "maximum": 26214400, "readOnly": true}, "media_type": {"type": "string", "enum": ["application/pdf", "application/vnd.openxmlformats-officedocument.wordprocessingml.document", "application/msword", "text/markdown; charset=utf-8", "text/plain; charset=utf-8"], "readOnly": true}, "status": {"type": "string", "const": "READY", "readOnly": true}}},
      "ResourceProjection": {"type": "object", "required": ["progress", "tags", "original_file", "review_plan"], "properties": {"progress": {"$ref": "#/components/schemas/LearningProgress"}, "tags": {"type": "array", "maxItems": 20, "items": {"$ref": "#/components/schemas/Tag"}}, "original_file": {"oneOf": [{"$ref": "#/components/schemas/ReadyOriginalFile"}, {"type": "null"}]}, "review_plan": {"oneOf": [{"$ref": "#/components/schemas/ActiveReviewPlan"}, {"type": "null"}]}}},
      "ResourceSummaryProjection": {"type": "object", "required": ["progress", "tags", "review_plan"], "properties": {"progress": {"$ref": "#/components/schemas/LearningProgress"}, "tags": {"type": "array", "maxItems": 20, "items": {"$ref": "#/components/schemas/Tag"}}, "review_plan": {"oneOf": [{"$ref": "#/components/schemas/ActiveReviewPlan"}, {"type": "null"}]}}},
      "WebResourceSummary": {"allOf": [{"$ref": "#/components/schemas/LearningResource"}, {"$ref": "#/components/schemas/ResourceSummaryProjection"}, {"type": "object", "required": ["source_type", "original_file"], "properties": {"source_type": {"const": "WEB"}, "original_file": {"type": "null"}}}], "unevaluatedProperties": false},
      "PasteResourceSummary": {"allOf": [{"$ref": "#/components/schemas/LearningResource"}, {"$ref": "#/components/schemas/ResourceSummaryProjection"}, {"type": "object", "required": ["source_type", "original_file"], "properties": {"source_type": {"const": "PASTE"}, "original_file": {"type": "null"}}}], "unevaluatedProperties": false},
      "FileResourceSummary": {"allOf": [{"$ref": "#/components/schemas/LearningResource"}, {"$ref": "#/components/schemas/ResourceSummaryProjection"}, {"type": "object", "required": ["source_type", "original_file"], "properties": {"source_type": {"const": "FILE"}, "original_file": {"$ref": "#/components/schemas/OriginalFileSummary"}}}], "unevaluatedProperties": false},
      "ResourceSummary": {"description": "按来源判别的列表安全投影；永远不含 source_url、pasted_content 或 sha256。", "oneOf": [{"$ref": "#/components/schemas/WebResourceSummary"}, {"$ref": "#/components/schemas/PasteResourceSummary"}, {"$ref": "#/components/schemas/FileResourceSummary"}], "discriminator": {"propertyName": "source_type", "mapping": {"WEB": "#/components/schemas/WebResourceSummary", "PASTE": "#/components/schemas/PasteResourceSummary", "FILE": "#/components/schemas/FileResourceSummary"}}},
      "WebResourceDetail": {"allOf": [{"$ref": "#/components/schemas/LearningResource"}, {"$ref": "#/components/schemas/ResourceProjection"}, {"type": "object", "required": ["source_type", "source_url"], "properties": {"source_type": {"const": "WEB"}, "source_url": {"$ref": "#/components/schemas/SourceUrl"}, "original_file": {"type": "null"}}}], "unevaluatedProperties": false},
      "PasteResourceDetail": {"allOf": [{"$ref": "#/components/schemas/LearningResource"}, {"$ref": "#/components/schemas/ResourceProjection"}, {"type": "object", "required": ["source_type", "pasted_content"], "properties": {"source_type": {"const": "PASTE"}, "pasted_content": {"type": "string", "minLength": 1, "maxLength": 1000000, "x-sensitive": true}, "original_file": {"type": "null"}}}], "unevaluatedProperties": false},
      "FileResourceDetail": {"allOf": [{"$ref": "#/components/schemas/LearningResource"}, {"$ref": "#/components/schemas/ResourceProjection"}, {"type": "object", "required": ["source_type", "original_file"], "properties": {"source_type": {"const": "FILE"}, "original_file": {"$ref": "#/components/schemas/ReadyOriginalFile"}}}], "unevaluatedProperties": false},
      "ResourceDetail": {"oneOf": [{"$ref": "#/components/schemas/WebResourceDetail"}, {"$ref": "#/components/schemas/PasteResourceDetail"}, {"$ref": "#/components/schemas/FileResourceDetail"}], "discriminator": {"propertyName": "source_type", "mapping": {"WEB": "#/components/schemas/WebResourceDetail", "PASTE": "#/components/schemas/PasteResourceDetail", "FILE": "#/components/schemas/FileResourceDetail"}}},
      "WebResourceCreate": {"type": "object", "additionalProperties": false, "required": ["source_type", "title", "source_url"], "properties": {"source_type": {"type": "string", "const": "WEB"}, "title": {"type": "string", "minLength": 1, "maxLength": 200}, "source_url": {"$ref": "#/components/schemas/SourceUrl"}, "source_name": {"type": ["string", "null"], "maxLength": 120}, "save_reason": {"type": ["string", "null"], "maxLength": 1000}, "topic_id": {"type": ["string", "null"], "format": "uuid"}, "tag_ids": {"type": "array", "maxItems": 20, "uniqueItems": true, "items": {"type": "string", "format": "uuid"}}}},
      "PasteResourceCreate": {"type": "object", "additionalProperties": false, "required": ["source_type", "title", "pasted_content"], "properties": {"source_type": {"type": "string", "const": "PASTE"}, "title": {"type": "string", "minLength": 1, "maxLength": 200}, "pasted_content": {"type": "string", "minLength": 1, "maxLength": 1000000, "x-sensitive": true}, "source_name": {"type": ["string", "null"], "maxLength": 120}, "save_reason": {"type": ["string", "null"], "maxLength": 1000}, "topic_id": {"type": ["string", "null"], "format": "uuid"}, "tag_ids": {"type": "array", "maxItems": 20, "uniqueItems": true, "items": {"type": "string", "format": "uuid"}}}},
      "FileResourceCreate": {"type": "object", "additionalProperties": false, "required": ["source_type", "title", "file"], "properties": {"source_type": {"type": "string", "const": "FILE"}, "title": {"type": "string", "minLength": 1, "maxLength": 200}, "file": {"type": "string", "format": "binary"}, "source_name": {"type": ["string", "null"], "maxLength": 120}, "save_reason": {"type": ["string", "null"], "maxLength": 1000}, "topic_id": {"type": ["string", "null"], "format": "uuid"}, "tag_ids": {"type": "array", "maxItems": 20, "uniqueItems": true, "items": {"type": "string", "format": "uuid"}}}},
      "ResourcePatch": {"type": "object", "additionalProperties": false, "minProperties": 2, "required": ["expected_version"], "properties": {"expected_version": {"type": "integer", "minimum": 1}, "title": {"type": "string", "minLength": 1, "maxLength": 200}, "source_url": {"$ref": "#/components/schemas/SourceUrl"}, "pasted_content": {"type": "string", "minLength": 1, "maxLength": 1000000, "x-sensitive": true}, "source_name": {"type": ["string", "null"], "maxLength": 120}, "save_reason": {"type": ["string", "null"], "maxLength": 1000}, "topic_id": {"type": ["string", "null"], "format": "uuid"}}, "not": {"required": ["source_url", "pasted_content"]}},
...
      "ResourceEnvelope": {"type": "object", "additionalProperties": false, "required": ["data"], "properties": {"data": {"$ref": "#/components/schemas/ResourceDetail"}}},
...
      "ResourcePage": {"type": "object", "additionalProperties": false, "required": ["data", "page"], "properties": {"data": {"type": "array", "items": {"$ref": "#/components/schemas/ResourceSummary"}}, "page": {"$ref": "#/components/schemas/PageMeta"}}},
...
    },
    "responses": {
      "BadRequest": {"description": "请求语法无法解析。", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/ErrorResponse"}, "example": {"error": {"code": "MALFORMED_REQUEST", "message": "请求格式不正确。", "details": {}, "request_id": "req_7cc640f6b1764ed0"}}}}},
...
      "ForbiddenRead": {"description": "Host 或本地令牌校验失败；读取操作不要求 Origin。", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/ErrorResponse"}, "examples": {"badHost": {"value": {"error": {"code": "HOST_FORBIDDEN", "message": "请求主机不受信任。", "details": {}, "request_id": "req_7cc640f6b1764ed0"}}}, "missingToken": {"value": {"error": {"code": "LOCAL_TOKEN_REQUIRED", "message": "需要当前运行会话。", "details": {}, "request_id": "req_7cc640f6b1764ed0"}}}, "invalidToken": {"value": {"error": {"code": "LOCAL_TOKEN_INVALID", "message": "当前运行会话无效。", "details": {}, "request_id": "req_7cc640f6b1764ed0"}}}}}}},
      "ForbiddenWrite": {"description": "Host、本地令牌或写请求来源校验失败；在读取请求体、数据库或文件前拒绝。", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/ErrorResponse"}, "examples": {"badHost": {"value": {"error": {"code": "HOST_FORBIDDEN", "message": "请求主机不受信任。", "details": {}, "request_id": "req_7cc640f6b1764ed0"}}}, "missingToken": {"value": {"error": {"code": "LOCAL_TOKEN_REQUIRED", "message": "需要当前运行会话。", "details": {}, "request_id": "req_7cc640f6b1764ed0"}}}, "invalidToken": {"value": {"error": {"code": "LOCAL_TOKEN_INVALID", "message": "当前运行会话无效。", "details": {}, "request_id": "req_7cc640f6b1764ed0"}}}, "badOrigin": {"value": {"error": {"code": "REQUEST_ORIGIN_FORBIDDEN", "message": "请求来源不受信任。", "details": {}, "request_id": "req_7cc640f6b1764ed0"}}}}}}},
      "ResourceNotFound": {"description": "资料不存在。", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/ErrorResponse"}, "examples": {"resource": {"value": {"error": {"code": "RESOURCE_NOT_FOUND", "message": "没有找到请求的资料。", "details": {}, "request_id": "req_7cc640f6b1764ed0"}}}}}}},
...
      "TopicOrTagNotFound": {"description": "创建资料引用的主题或标签不存在。", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/ErrorResponse"}, "examples": {"topic": {"value": {"error": {"code": "TOPIC_NOT_FOUND", "message": "没有找到引用的主题。", "details": {}, "request_id": "req_7cc640f6b1764ed0"}}}, "tag": {"value": {"error": {"code": "TAG_NOT_FOUND", "message": "没有找到引用的标签。", "details": {}, "request_id": "req_7cc640f6b1764ed0"}}}}}}},
      "ResourceOrTopicNotFound": {"description": "资料或其引用的主题不存在。", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/ErrorResponse"}, "examples": {"resource": {"value": {"error": {"code": "RESOURCE_NOT_FOUND", "message": "没有找到请求的资料。", "details": {}, "request_id": "req_7cc640f6b1764ed0"}}}, "topic": {"value": {"error": {"code": "TOPIC_NOT_FOUND", "message": "没有找到引用的主题。", "details": {}, "request_id": "req_7cc640f6b1764ed0"}}}}}}},
...
      "ResourceUpdateConflict": {"description": "资料版本或来源字段冲突。", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/ErrorResponse"}, "examples": {"version": {"value": {"error": {"code": "VERSION_CONFLICT", "message": "数据已经发生变化，请刷新后重试。", "details": {"current_version": 4}, "request_id": "req_7cc640f6b1764ed0"}}}, "sourceType": {"value": {"error": {"code": "SOURCE_TYPE_MISMATCH", "message": "来源字段与资料类型不匹配。", "details": {}, "request_id": "req_7cc640f6b1764ed0"}}}}}}},
...
      "Gone": {"description": "删除确认令牌已过期。", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/ErrorResponse"}, "example": {"error": {"code": "DELETION_TOKEN_EXPIRED", "message": "删除确认已经过期，请重新预览。", "details": {}, "request_id": "req_7cc640f6b1764ed0"}}}}},
      "TooLarge": {"description": "文件超过 26,214,400 字节。", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/ErrorResponse"}, "example": {"error": {"code": "FILE_TOO_LARGE", "message": "文件超过 25 MiB 上限。", "details": {"max_bytes": 26214400}, "request_id": "req_7cc640f6b1764ed0"}}}}},
      "UnsupportedMedia": {"description": "请求媒体类型或文件格式不支持。", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/ErrorResponse"}, "examples": {"contentType": {"value": {"error": {"code": "CONTENT_TYPE_UNSUPPORTED", "message": "请求内容类型不受支持。", "details": {}, "request_id": "req_7cc640f6b1764ed0"}}}, "fileType": {"value": {"error": {"code": "FILE_TYPE_UNSUPPORTED", "message": "文件格式不受支持或内容与格式不一致。", "details": {}, "request_id": "req_7cc640f6b1764ed0"}}}}}}},
      "Validation": {"description": "字段、范围、互斥、筛选或排序无效。", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/ErrorResponse"}, "example": {"error": {"code": "VALIDATION_ERROR", "message": "请检查填写内容。", "details": {"fields": [{"field": "title", "reason": "不能为空"}]}, "request_id": "req_7cc640f6b1764ed0"}}}}},
      "PreconditionRequired": {"description": "缺少版本前置条件：带 JSON 请求体的修改操作缺少 expected_version，或无请求体的删除操作缺少 If-Match。", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/ErrorResponse"}, "example": {"error": {"code": "VERSION_REQUIRED", "message": "请刷新后再执行此操作。", "details": {}, "request_id": "req_7cc640f6b1764ed0"}}}}},
      "ServerError": {"description": "未知服务错误。", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/ErrorResponse"}, "example": {"error": {"code": "UNKNOWN_ERROR", "message": "服务暂时无法完成请求。", "details": {}, "request_id": "req_7cc640f6b1764ed0"}}}}},
      "StorageUnavailable": {"description": "受控存储路径不可安全访问。", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/ErrorResponse"}, "example": {"error": {"code": "STORAGE_PATH_UNAVAILABLE", "message": "原始文件存储暂时不可用。", "details": {}, "request_id": "req_7cc640f6b1764ed0"}}}}},
      "UnsupportedContentType": {"description": "请求 Content-Type 不是该操作声明的媒体类型。", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/ErrorResponse"}, "example": {"error": {"code": "CONTENT_TYPE_UNSUPPORTED", "message": "请求内容类型不受支持。", "details": {}, "request_id": "req_7cc640f6b1764ed0"}}}}}
    }
  }
}